move rest-client to own module
