+++ msa/black-box-tests/pom.xml
@@ -90,6 +90,10 @@
             <groupId>org.thingsboard</groupId>
             <artifactId>tools</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.thingsboard</groupId>
+            <artifactId>rest-client</artifactId>
+        </dependency>
     </dependencies>
 
     <build>
